Port 17 functions from render_cameras.obj
Functions ported:
- render_camera_build_frustum_bounds @ 0x185950 (render_cameras.obj) [100.0% VC71, 97/97 insns]
- render_frustum_get_projection_bounds @ 0x185f80 (render_cameras.obj) [100.0% VC71, 58/58 insns]
- render_camera_screen_to_view @ 0x186230 (render_cameras.obj) [100.0% VC71, 75/75 insns]
- render_camera_triangle_frontfacing @ 0x1865e0 (render_cameras.obj) [94.0% VC71, 68/66 insns]
- render_frustum_build_point_flags @ 0x186690 (render_cameras.obj) [100.0% VC71, 78/78 insns]
- render_camera_view_to_screen @ 0x186050 (render_cameras.obj) [100.0% VC71, 145/145 insns]
- render_camera_build_clipped_frustum_bounds @ 0x186480 (render_cameras.obj) [99.1% VC71, 116/116 insns]
- render_frustum_sphere_visible @ 0x186ac0 (render_cameras.obj) [93.7% VC71, 143/143 insns]
- render_frustum_cube_view_fraction @ 0x185ad0 (render_cameras.obj) [97.1% VC71, 365/358 insns]
- render_frustum_cube_visible @ 0x1867f0 (render_cameras.obj) [98.4% VC71, 243/243 insns]
- render_camera_debug_frustum @ 0x186d40 (render_cameras.obj) [98.1% VC71, 135/134 insns]
- render_camera_mirror @ 0x186ef0 (render_cameras.obj) [94.8% VC71, 306/304 insns]
- contrail_fade @ 0x187f80 (render_cameras.obj) [100.0% VC71, 55/55 insns]
- render_contrail @ 0x188010 (render_cameras.obj) [99.5% VC71, 634/634 insns]

* Adds real_rectangle3d (0x18) to types.h; render_frustum_cube_visible now takes (float *bounds, bool test_frustum_against_cube) - callers cast only.
* Drops 12 stale FUN_000a* floor rows in weather_particle_systems.c that duplicate their renamed entries (renamed in b14b19d0f).
* Adds render_mirror_t (plane/index_of_refraction/depth, proven offsets only).
* Adds real_argb_color, contrail_datum_t, contrail_point_datum_t, contrail_point_state_t, contrail_shader_t
and contrail_definition_t (proven offsets only; PAL 2342 names).  render_type jump table read from 0x188790: 0 vertical, 1/2 horizontal/media, 4 viewer, 3 = error.

Coverage: 52.7% (4924/9348 symbols)

+++ src/halo/effects/weather_particle_systems.c
@@ -873,7 +873,7 @@ void weather_particle_system_render(int16_t local_player_index /* @<eax> */)
           cube[3] = tile_offset[j] + box_max_y;
           cube[5] = tile_offset[k] + box_max_z;
 
-          if (render_frustum_cube_visible((void *)0x5065a4, (int)(size_t)cube,
+          if (render_frustum_cube_visible((void *)0x5065a4, cube,
                                           1) != 0) {
             assert_halt_msg_at(
               "box_count<MAXIMUM_NUMBER_OF_VISIBLE_WEATHER_PARTICLE_BOXES",
